pyproject.toml
[project]
name = "guide_creator_flow"
version = "0.1.0"
description = "guide_creator_flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.203.1,<1.0.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.scripts]
kickoff = "guide_creator_flow.main:kickoff"
run_crew = "guide_creator_flow.main:kickoff"
plot = "guide_creator_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
